rename chart to be more explicit
