+++ src/app/js/formats/contextual-bar-chart/AdminComponent.js
