+++ src/app/js/formats/contextual-bar-chart/Component.js
