+++ src/app/js/formats/contextual-bar-chart/index.js
